Move python compatability code into compat submodule.
